Restructured \Core\Validator
